+++ lib/src/upgrade_messages.dart
@@ -55,8 +55,8 @@ class UpgraderMessages {
   final String languageCode;
 
   /// Provide a [code] to override the system-reported default locale.
-  UpgraderMessages({String? code})
-      : languageCode = (code ?? findLanguageCode()) {
+  UpgraderMessages({String? code, BuildContext? context})
+      : languageCode = (code ?? findLanguageCode(context: context)) {
     assert(languageCode.isNotEmpty);
   }