+++ example/lib/main.dart
@@ -19,6 +19,10 @@ void main() async {
   runApp(const MyApp());
 }
 
+final Upgrader upgrader = Upgrader(
+  debugLogging: true,
+);
+
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
@@ -27,6 +31,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Upgrader Example',
       home: UpgradeAlert(
+        upgrader: upgrader,
         child: Scaffold(
           appBar: AppBar(title: const Text('Upgrader Example')),
           body: const Center(child: Text('Checking...')),